requirements.txt
openai==1.5.0
streamlit==1.29.0
langchain==0.0.352
bs4==0.0.1
tiktoken==0.5.2
python-dotenv==1.0.0
langchain-experimental==0.0.47
duckduckgo-search==4.1.1
numexpr==2.8.8
requests==2.31.0